Version 1.1.3
- Bug Fix of reload command and next command (for paper users);
- Added Compatibility with 1.19.4;
